+++ src/integrations/react.js
@@ -1,11 +1,16 @@
 import { createRoot } from 'react-dom/client';
 
-export default ({ name, module, props, element }) => {
-    const component = module[name] ?? module.default;
+export default (promise) => {
+    return async ({ name, props, element }) => {
+        const module = await promise;
+        const component = module[name] ?? module.default;
 
-    return new Promise((resolve) => {
-        const app = createRoot(element);
-        app.render(component({ ...props }));
-        resolve(app);
-    });
+        return new Promise((resolve) => {
+            const app = createRoot(element);
+            app.render(component({ ...props }));
+            resolve({
+                root: app
+            });
+        });
+    };
 };