+++ src/integrations/vue.js
@@ -1,19 +1,24 @@
 import { createApp, h } from 'vue';
 
-export default ({ name, module, props, element }) => {
-    const component = module.default;
+export default (promise) => {
+    return async ({ name, props, element }) => {
+        const module = await promise;
+        const component = module.default;
 
-    return new Promise((resolve) => {
-        const app = createApp({
-            name: component.name + '-host',
-            mounted() {
-                resolve(app);
-            },
-            render() {
-                return h(component, props);
-            }
+        return new Promise((resolve) => {
+            const app = createApp({
+                name: component.name + '-host',
+                mounted() {
+                    resolve({
+                        root: app
+                    });
+                },
+                render() {
+                    return h(component, props);
+                }
+            });
+
+            app.mount(element);
         });
-
-        app.mount(element);
-    });
-};
+    };
+}