+++ src/navigation.ts
@@ -6,11 +6,15 @@ import PageIsland from "./components/page-island";
 import replace from './page-replace';
 import FragmentIsland from "./components/fragment-island";
 
+type AdaptiveNavigateEvent = NavigateEvent & {
+    originalEvent?: MouseEvent | SubmitEvent;
+};
+
 export const setupNavigation = () => {
-    window.navigation?.addEventListener('navigate', (event: NavigateEvent) => {
+    window.navigation?.addEventListener('navigate', (event: AdaptiveNavigateEvent) => {
         navigate(
             event.destination.url,
-            event.target as HTMLElement | null,
+            (event.sourceElement ?? event.originalEvent?.target) as HTMLElement | null,
             event.formData,
             event
         );
@@ -81,12 +85,7 @@ const navigate = (
         return;
     }
 
-    let method: string = 'GET';
-
-    if (formData) {
-        method = 'POST';
-    }
-
+    targetElement?.classList.add('target-loading');
     let frame: FragmentIsland | PageIsland | null = null;
 
     if (
@@ -97,11 +96,9 @@ const navigate = (
         targetElement
     ) {
         frame = targetElement.closest('page-island, fragment-island');
-        console.log('frame', frame);
 
         if (target === '_parent' && frame) {
             frame = frame.parentElement?.closest('page-island, fragment-island') ?? null;
-            console.log('frame2', frame);
         }
     } else if (target) {
         frame = document.querySelector(`fragment-island[name=${target}]`);
@@ -116,36 +113,53 @@ const navigate = (
         return;
     }
 
+    const handler = () => loadDestination({
+        url,
+        frame,
+        targetElement,
+        method: formData ? 'POST' : 'GET',
+        formData,
+    }).then(() => {
+        targetElement?.classList.remove('target-loading');
+    });
+
 
     if (navigateEvent) {
         navigateEvent.intercept({
             focusReset: formData ? 'manual' : 'after-transition',
             scroll: formData ? 'manual' : 'after-transition',
-            handler: () => loadDestination(
-                frame,
-                url,
-                method,
-                formData,
-            )
+            handler: handler
         });
     } else {
-        loadDestination(
-            frame,
-            url,
-            method,
-            formData
-        );
+        handler();
     }
 };
 
-export const loadDestination = async (
-    frame: FragmentIsland | PageIsland,
+let abortController: AbortController;
+
+export const loadDestination = async ({
+    url, frame, targetElement, method, formData
+}: {
     url: string,
+    frame: FragmentIsland | PageIsland,
+    targetElement?: HTMLElement | null,
     method: string,
-    formData: FormData | null = null
-) => {
+    formData?: FormData | null
+}): Promise<string> => {
+    if (abortController) {
+        abortController.abort();
+    }
+
+    abortController = new AbortController();
     const headers = {};
     frame._onLoadStart();
+
+    const templateId = targetElement?.getAttribute('loading') || frame.getAttribute('loading');
+    const template: HTMLTemplateElement | null = templateId ? document.querySelector(`template[id="${templateId}"]`) : null;
+
+    if (template) {
+        replace(frame, template.innerHTML);
+    }
 
     if (frame.tagName === 'PAGE-ISLAND') {
         headers['X-Page-Island'] = 'page';
@@ -153,14 +167,16 @@ export const loadDestination = async (
         headers['X-Fragment-Island'] = frame.getAttribute('name') || 'default';
     }
 
-    fetch(url, {
+    return fetch(url, {
         method,
         body: formData,
-        headers
+        headers,
+        signal: abortController.signal,
     })
         .then(response => response.text())
         .then(data => {
             replace(frame, data);
             frame._onLoadEnd();
+            return data;
         });
 }