+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v0.1.0 (2025-03-?)
 * Initial component implementation
 * Added prop support
+* Built initial Page island structure